Load services from configDirectory
Grunt looks for files in configDirectory and loads any service
definitions it finds.  The configDirectory option in the gruntfile.yml
specifies a directory that is walked.  Any non-directory is loaded
as a YML file that ought to contain only service definitions.  This
makes it much easier to add services to the Docker.